+++ packages/frontend/src/components/LoginPage/index.ts
@@ -1,1 +1,1 @@
-export { useVkFloatingOneTap } from "./LoginPage";
+export { useVkAuth } from "./LoginPage";